frankenshtain

+++ lesson_8.cashingproxy/src/main/java/lesson_8/cashingproxy/calulator/pac/ServiceImpl.java
@@ -3,6 +3,6 @@ package lesson_8.cashingproxy.calulator.pac;
 public class ServiceImpl implements Service {
     @Override
     public double doHardWork(String s, Integer i) {
-        return s.length()/i;
+        return (double) s.length()/i;
     }
 }
